requirements.txt
pywebview>=4.0
pygments>=2.10.0